+++ edx_proctoring/serializers.py
@@ -96,7 +96,7 @@ class ProctoredExamStudentAttemptSerializer(serializers.ModelSerializer):
             "external_id", "status", "proctored_exam", "allowed_time_limit_mins",
             "attempt_code", "is_sample_attempt", "taking_as_proctored",
             "review_policy_id", "is_status_acknowledged",
-            "time_remaining_seconds", "is_resumable"
+            "time_remaining_seconds", "is_resumable", "ready_to_resume", "resumed"
         )